Extract each offer class into a file of its own
