+++ csharp/SupermarketReceipt/Offer.cs
@@ -40,68 +40,4 @@ namespace SupermarketReceipt
         public abstract Discount CalculateDiscount(double unitPrice, double totalQuantityForProduct);
 
     }
-
-    public class FivePerAmountOffer : Offer
-    {
-        public FivePerAmountOffer(Product product, double argument) : base(product, argument)
-        {
-        }
-        
-        public override Discount CalculateDiscount(double unitPrice, double quantity)
-        {
-            var quantityAsInt = (int)quantity;
-            if (quantityAsInt < 5)
-                return null;
-            var numberOfDiscountsToApply = quantityAsInt / 5;
-            var discountTotal = unitPrice * quantity - (Argument * numberOfDiscountsToApply + quantityAsInt % 5 * unitPrice);
-            return new Discount(_product, 5 + " for " + Argument, -discountTotal);
-        }
-    }
-
-    public class TenPercentDiscountOffer : Offer
-    {
-        public TenPercentDiscountOffer(Product product, double argument) : base(product, argument)
-        {
-        }
-        
-        public override Discount CalculateDiscount(double unitPrice, double quantity)
-        {
-            return new Discount(_product, Argument + "% off", -quantity * unitPrice * Argument / 100.0);
-        }
-    }
-
-    public class ThreeForTwoDiscount : Offer
-    {
-        public ThreeForTwoDiscount(Product product, double argument) : base(product, argument)
-        {
-        }
-        public override Discount CalculateDiscount(double unitPrice, double quantity)
-        {
-            var quantityAsInt = (int)quantity;
-            if (quantityAsInt < 3)
-                return null;
-            var numberOfDiscountsToApply = quantityAsInt / 3;
-            var discountAmount = quantity * unitPrice - (numberOfDiscountsToApply * 2 * unitPrice + quantityAsInt % 3 * unitPrice);
-            return new Discount(_product, "3 for 2", -discountAmount);
-        }
-    }
-
-    public class TwoForAmountOffer : Offer
-    {
-        public TwoForAmountOffer(Product product, double argument) : base(product, argument)
-        {
-        }
-        
-        public override Discount CalculateDiscount(double unitPrice, double quantity)
-        {
-            var quantityAsInt = (int)quantity;
-            if (quantityAsInt < 2)
-                return null;
-            
-            var total = Argument * (quantityAsInt / 2) + quantityAsInt % 2 * unitPrice;
-            var discountN = unitPrice * quantity - total;
-            return new Discount(_product, "2 for " + Argument, -discountN);
-        }
-        
-    }
 }
